requirements.txt
flask==3.0.0
twilio==8.0.0
pandas==2.0.3
matplotlib==3.7.2
transformers==4.30.2
torch==2.0.1
psycopg2-binary==2.9.6
flask-talisman==1.0.0
python-dotenv==1.0.0